requirements.txt
cryptography==3.4.7
Pillow==8.3.1
tk==0.1.0